♻️ refactor: orderContext로 구현되던 orderProgress 분리
